Windows installer signing with code sign certificate
Windows installer executable should be signed to improve windows
user experience when installing the package. Unsigned installer
are causing the operating system to show a message saying that the
application may be dangerous and that it is unsafe to install
packages from unknown publishers. Signing the executable with a
valid code sign certificate will eliminate this behaviour.